requirements.txt
opencv-python~=4.11.0.86
numpy~=2.1.1
matplotlib~=3.10.1
torch~=2.8.0.dev20250321+cu128
pillow~=11.0.0
requests~=2.32.3
langgraph~=0.3.11
flask~=3.1.0
werkzeug~=3.1.3
typing_extensions~=4.12.2
pydantic~=2.10.6
torchvision~=0.22.0.dev20250322+cu128
rasterio~=1.4.3
scikit-learn~=1.6.1
scikit-image~=0.25.2
timm~=1.0.15
dotenv~=0.9.9
python-dotenv~=1.0.1
langchain-core~=0.3.45